# backend/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Friendo App API
  version: 1.0.0
  description: API Documentation for Friendo App Backend
servers:
  - url: http://localhost:3000
    description: Development Server
  - url: https://friendo-backend.vercel.app
    description: Production Server (Example)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/users/search:
    get:
      summary: Search for users
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: The search query to match against usernames
      responses:
        "200":
          description: List of matched users
        "401":
          description: Unauthorized
  /api/users/{uid}:
    get:
      summary: Get user profile by UID
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: string
          description: The user ID
      responses:
        "200":
          description: User profile
        "404":
          description: User not found
  /api/users/me:
    put:
      summary: Update current user profile
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                photoURL:
                  type: string
                bio:
                  type: string
                dob:
                  type: string
                  format: date
      responses:
        "200":
          description: Successfully updated profile
        "401":
          description: Unauthorized
...